// server/src/schemas/authSchema.ts
import { z } from "zod";

export const registerSchema = z.object({
    name: z.string().min(2, "Name required"),
    email: z.string().email("Invalid email"),
    role: z.enum(["Admin", "Sales User"]),
    password: z.string().min(6, "Password must be 6+ chars")
});

export const loginSchema = z.object({
    email: z.string().email("Invalid email"),
    password: z.string().min(6, "Password must be 6+ chars")
});